Crate js_connectors

Query Engine JS Connectors This crate is responsible for defining a quaint::Connector implementation that uses functions exposed by client connectors via N-API.

A JsConnector is an object defined in javascript that uses a driver (ex. ‘@planetscale/database’) to provide a similar implementation of that of a quaint Connector. i.e. the ability to query and execute SQL plus some transformation of types to adhere to what a quaint::Value expresses.

Structs

JsQueryable

A JsQueryable adapts a Proxy to implement quaint’s Queryable interface. It has the responsibility of transforming inputs and outputs of query and execute methods from quaint types to types that can be translated into javascript and viceversa. This is to let the rest of the query engine work as if it was using quaint itself. The aforementioned transformations are:

Functions

from_napi